Added ability to zoom in on clicked photo
I added the ability to zoom in on the photo that you click. I found that
there was a weird thing happening with the order of execution for
loading photo info that was causing photos to be out of order. in order
to align this with what is displayed I moved the creating of the image
tags to the photo info loading method.